int,int -> size_t,size_t

+++ tasks/kulik_a_the_most_different_adjacent/mpi/src/ops_mpi.cpp
@@ -55,7 +55,7 @@ bool KulikATheMostDifferentAdjacentMPI::RunImpl() {
   }
   struct {
     double val;
-    int ind;
+    size_t ind;
   } max_diff, max_diffall;
   max_diff.val = max_diffall.val = 0.;
   max_diff.ind = max_diffall.ind = 0;
